Increase tolerance_num of Cpu RNNLayer tests
Instead of increasing the tolerance amount, we increase the number of elements we tolerate to 0.02 % of the whole tensor. This ensures we do not affect the tolerance for smaller tests.

This amount is set according to the number of elements above the threshold. it was 1 over 512 elements, 1/512 ~ 0.02 %.

Resolves: COMPMID-6932

Change-Id: I9d3ce29a3972aa8b9daea5288005a0a41a266328
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/11321
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Pablo Marquez Tello <pablo.tello@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>