Waive overflow issue in a64_gemm_s8_4x4
a64_gemm_s8_4x4 assembly kernel has an overflow issue, if fixed, could case the kernel to perform much slower.

For this to happen, Lhs matrix must have -128 values eight positions apart and the Rhs matrix must have -128 in the same positions as well. So, the multiplication becomes (-128)*(-128) + (-128)*(-128) = 32768, which overflows and results in being -32768.

The possibility of this happening is very low and when it happens, it's going to affect only one pixel. Therefore, we waive the issue and report it in the errata. We also modify the relevant test with minimum Int8 to use as -127 instead of -128.

Change-Id: Ia36407d67c439eb14c145aede2f07729bc41db2e
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/11818
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Omar Al Khatib <omar.alkhatib@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>