docs: Put only the public interface in the documentation
The documentation of ComputeLibrary includes both public interface and the implementation classes/files. Having files/classes that are not exposed to the user in the documentation bloats the size and makes it more difficult to release and clone.

It also removes internal references, mentioned by @ref, in the comment headers to prevent errors and warnings.

Resolves: COMPMID-7375

Change-Id: I0d58460faebdacbd478e9d38b20d62a77430b656
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12107
Reviewed-by: Ramy Elgammal <ramy.elgammal@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>