docs: Update the documentation links
We also revert using SHORT_NAMES option in Doxyfile, that was enabled in this patch:

https://review.mlplatform.org/c/ml/ComputeLibrary/+/12107

Because, we refer to some files by their explicit names in our documentation. Using short names will make those links invalid.

We change all 'latest' documentation references to the specific release version so that if people go back in time to a past tag, they can access that versions' documentation by just clicking the Documentation button. This is not the case at the moment, which is confusing.

We keep the one in CONTRIBUTING.md because contributions should be based the latest guidelines.

Change-Id: I367e0266beb384fd8298b91311a560661b688032
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12113
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Ramy Elgammal <ramy.elgammal@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>