fix: Validation tests compute reference and target on each iteration
Validation tests that are derived from Fixture class are written in a
way that they compute reference and target outputs in their setup()
methods, and their run() method compares pre-computed values only. This
is not what is expected when the user runs validation tests with a
number of iterations (--iterations flag). The expected behavior is to
perform computation on each iteration.

Benchmarks that are derived from Fixture on the other hand behave
correctly - their setup() method is expected to be executed once, and
they do calculations in their run() methods on each iteration.

The patch fixes the behavior for validation tests and leaves the
benchmarks unaffected.

Resolves: COMPMID-7430
Signed-off-by: Yevgen Pronenko <yevgen.pronenko@arm.com>
Change-Id: I47b9cdd947df2746933e912e7351f830fc0ae344
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12201
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Gunes Bayir <gunes.bayir@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>