fix: rounding of negative integers in pooling 2d/3d kernels
The pool3d and the non-assembly pool2d kernels were rounding the result by adding 0.5f and then casting to int. This doesn't work well with negative values. When the result is, say, -5, if we add 0.5 and cast to integer it becomes -4, where it should have stayed as -5.

This patch fixes this by properly rounding the values using "rounding to nearest with ties to even" policy. This is the standard rounding policy that other ML frameworks are using and is said to result in less bias compared to std::round().

If it's 32-bit build, the corresponding vector instructions do not exist. Therefore, rounding towards zero policy is used.

Resolves: COMPMID-7441
Change-Id: I1e191b9592e646e1fff7ec769ccad238a03c7b0c
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12279
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Ramy Elgammal <ramy.elgammal@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>