fix: Incorrect quantized data casting in Gpu
Cast kernel in Gpu backend performs a specific operation in quantized mode, but this behavior was added only for GEMMLowp in Gpu backend to convert QSYMM8_PER_CHANNEL to QASYMM8.

Cast operation does not quantize/dequantize or requantize between quantized data types. Users should use NEQuantizationLayer or NEDequantizationLayer for this purpose. Casting is for the underlying data types. For example, QASYMM8 will be treated as if it is U8. QSYMM8_PER_CHANNEL's underlying data type is Int8 (S8). So, the use case in GEMMLowp is actually U8 to S8, but it needs a different conversion between these two types (XOR with 80h). For this reason, this patch adds a check to the public interface function's validate method, so that users do not try to perform this conversion, but the kernel can still do it in favor of continuing supporting GEMMLowp in Gpu backend. All the other conversions between quantized types are valid as long as they are not the same type.

Examples:
    QSYMM8_PER_CHANNEL --> QSYMM16 (Int8 to Int16)
    QASYMM8 --> QSYMM8_PER_CHANNEL (UInt8 to Int8)
    QASYMM8_SIGNED --> QSYMM8_PER_CHANNEL (Not Valid because they have the same underlying data type)
    QSYMM8_PER_CHANNEL --> QASYMM8 (Not Valid, reserved for GEMMLowp at the kernel level)
    QASYMM16 --> QSYMM8_PER_CHANNEL --> (UInt16 --> Int8)

All of the type checks have been added as Validate test cases, and the headers are updated.

Resolves: COMPMID-7439

Change-Id: Id94299e6ca5dd938dcfe88cd7e1c9f658d1e2730
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12414
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Ramy Elgammal <ramy.elgammal@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>