fix: make cpu cast truncate float to int
Float to Int cast was not consistent with other ML frameworks' casting behavior, and also C++ static_cast.  So, the float number will be truncated instead of rounding to the nearest integer.

For example, 1.99 will be casted to 1 (before it was 2).

Resolves: COMPMID-7446

Change-Id: Ie318b57c9eb5dc88a069ad40aca95a9fb41467b1
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12412
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Andreas Flöjt <andreas.floejt@arm.com>