fix: Apply consistent rounding policy in NEReduceMean
The rounding behavior in this operator was not consistent, sometimes truncating, sometimes rounding to the closest integer. The truncation usually should not be used because it'll cause a value like 1.999 to become 1, which is not a good approximation. Also, the rounding policy while converting from float to int has been changed to Round Closest to Integer with Ties to Nearest Even; instead of Round to Closest with Ties Away From Zero because it'll have less bias.

An Example: (139 + 154 + 164 + 179) / 4 is 159 w/o the need of rounding. However, if we have to do this operation recursively/repetitively, i.e.

```
mean( mean(139, 154), mean(164, 179) )
```
The results are going to be different depending on the rounding policy chosen.

With Ties Away From Zero:
```
mean(139, 154) = (139 + 154) / 2 = 146.5 --> 147
mean(164, 179) = (164 + 179) / 2 = 171.5 --> 172
mean(147, 172) = (147 + 172) / 2 = 159.5 --> 160
```
where we round up (away from zero) in both cases, causing an addition of 0.5 in both scenarios. And the final result changes by 1.

With Ties to Nearest Even:
```
mean(139, 154) = (139 + 154) / 2 = 146.5 --> 146
mean(164, 179) = (164 + 179) / 2 = 171.5 --> 172
mean(147, 172) = (146 + 172) / 2 = 159 --> 159
```
where we round down in the former and up in the latter, which balances the total, and we have the same result with the 4-element mean.

This occurs because when total, T, is divided into two pieces, X and T-X, the it's either Odd + Odd or Even + Even (if T is even as in our case). In the case where we require rounding, i.e. Odd+Odd, if X is rounded down, T-X will be rounded up because (T-X)/2 = T/2 (Integer) - X/2. And, X/2 and -X/2 will be rounded towards the same absolute value, i.e. if 61 / 2 --> 30 then -61/2 --> -30. In one, the closest even number is towards zero (i.e. 30.5 --> 30), so it's rounded down, and in the other one, it's rounded up (i.e. -30.5 --> -30). This balances to total sum.

This rounding mode has also been chosen by other ML frameworks for similar reasons.

Resolves: COMPMID-7496

Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Change-Id: I9173078e1678a6231b978a1f84f6c01ed3989428
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12489
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Sang Won Ha <sangwon.ha@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>