feat: add stateless gemm execution via ICPPKernel::run_op
This is a hot-fix of some gemm kernels to enable them to accept input
and output pointers at execution time. This allows them to be
thread-safe in inter-threaded environments. We expose this capability by
overriding ICPPKernel::run_op for CpuGemmAssemblyWrapperKernel. This
allows us to pass the kernel as well as the TensorPack to be operated
on.

Currently the following kernels are still stateful:
    * arm_gemm/gemm_hybrid_quantized
    * arm_gemm/gemm_interleaved
    * arm_gemm/gemv_batched
    * arm_gemm/quantize_wrapper

These currently fallback to the previous stateful execution mode where
arrays are read from the member data of arm_gemm::GemmCommon. Enabling
stateless execution for these kernels remains a future objective.

Change-Id: If72d702cc8af3a61a1b734f1b931ffd80c55df87
Signed-off-by: Siddhartha Menon <siddhartha.menon@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/12980
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Pablo Marquez Tello <pablo.tello@arm.com>
Reviewed-by: Gunes Bayir <gunes.bayir@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>