fix: Usages of dynamic shapes in the library
This is a BREAKING change.

This patch includes fixes, refactorings and additional tests required to provide a sensible implementation for dynamic shape support in the library.

We recently added dynamic shape support for TensorShape class, overloading the meaning of 0 value in the tensor dimension and have encountered a few edge cases where we need to interpret the value(s) of zero as an empty tensor or dynamic dimension. Although these cases can be overcome by workarounds, the best approach is to retract the implementation before causing more noise.

Edge cases encountered so far:
    - We interpret the tensor as "empty" if its number of dimensions is 0 (i.e. not initialized).
    - Reverse operator in some higher level frameworks work with empty axis tensor and output the Src as Dst, where in our case, this is interpreted as a dynamic shape and we validate to false.

To prevent additional edge cases appearing, and the existing dynamic shape implementation in the library through the ITensorInfo class will be adopted. But, certain tests will be added to to TensorInfo and SubTensorInfo classes to make sure a sensible implementation is in place.

Please also note that this feature was NOT fully functional and the tests associated with it in the Elementwise Binary and Unary were INCOMPLETE (i.e. not working as they were supposed to work), and were lacking validation. Thus, this patch also addresses those aspects. More info about the implementation details are here: COMPMID-2513

Because the feature is not fully correct, we implement this breaking change immediately, fixing it and making it available for ElementwiseUnary and Binary operators.

Implemented rules are
    - SubTensorInfo can't be dynamic, and can't have a dynamic parent because having a dynamic subtensor over a static one OR having a static subtensor over a dynamic one, along with pre-defined coordinates is not well-defined in the library.
    - TensorInfo can have dynamic shapes in a few of its arbitrary dimensions or all.

Tests and operators fixed:
    - NEElementwiseUnaryLayer and NEElementwiseOperations now support dynamic shapes
    - Tests were not stressing the case where the shapes were not available at configure time(). Fixtures are amended accordingly.
    -  Additional unit tests have been added for TensorInfo and SubTensorInfo classes.
    - Validations have been fixed to reflect this support level in the tests.

Partially Resolves: COMPMID-7676
Partially Resolves: COMPMID-7677
Resolves: COMPMID-7847

Change-Id: I044ca3443d42d49171a457d61acad803aae29753
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/13375
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Yevgen Pronenko <yevgen.pronenko@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>