fix: Apply -fPIC flag both to C and C++ code
This commit adds a new scons flag: test_static_dynamic_linking
This flag can be used by CI to verify that the Compute Library,
when linked as a static library, can itself be linked as a part
of a dynamic library.

This test ensures that all code in the Compute Library was
compiled as position independent code.

For this flag to be used, two conditions should be true:
- the library should be linked as static library (for instance,
due to bare-metal build, or standalone flag being set to 1.
- C library that is statically linked to the Compute Library
should also be compiled as position independent code.

Resolves: COMPMID-8067

Signed-off-by: Yevgen Pronenko <yevgen.pronenko@arm.com>
Change-Id: Id53234533898c26207887f4e06ac9eb011d51cdf
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/13740
Reviewed-by: Gunes Bayir <gunes.bayir@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>