refactor: Refactor reorder kernel and layer
Context: This is part of a new piece of work where we are looking to
enable more reorders through the arm_gemm::Transform kernels. In order
to do this, the following issues need to be fixed.

1.  The reorder kernel currently assumes the user knows what
blocking format will be used depending on the output data type and
has no control over whether the reorder will involve a transpose or not.
We need to introduce a more stringent validation where a whitelist of
configurations are listed.

2. We need to add a new parameter which decides whether
the data gets transposed or not.

- Refactor to use a map of allowed transform configurations
- Add "transpose" parameter
- Refactor the tests to use the new API

* No new test cases are added as no new formats have been enabled in this PR

Change-Id: I93edeb704961f20af550c5c0b313c588e4099433
Signed-off-by: Ryo Suzuki <ryo.suzuki@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/13762
Reviewed-by: Gunes Bayir <gunes.bayir@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>