fix: SME softmax Fp32 kernel failing given large inputs
Minimum value was not set correctly in the SME Fp32 Softmax kernel.
This does not cause problems if values are not too large, but it decreases
numerical stability. If values are large, it can produce NaN or Inf.

Resolves: COMPMID-8154

Change-Id: I55dab6a5b18279d8e4df0c2542d95c2d87bc9716
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/14180
Reviewed-by: Yevgen Pronenko <yevgen.pronenko@arm.com>
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Dennis Wildmark <dennis.wildmark@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>