fix: Mixed type quantized matmul for after configure quantization update
If the quantization information is required to be updated after configure, in some IPs, the mixed data type support does not return the correct results. This is because if there is no native mixed quantized kernel support in arm_gemm, the operator changes the sign of Lhs. This wasn't accounted correctly in the after configure update.

Resolves: COMPMID-7833
Change-Id: I88bb88921c569a9ef3b4aa49d09015f9e10247e8
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>
Reviewed-on: https://review.mlplatform.org/c/ml/ComputeLibrary/+/14670
Benchmark: Arm Jenkins <bsgcomp@arm.com>
Reviewed-by: Dennis Wildmark <dennis.wildmark@arm.com>
Comments-Addressed: Arm Jenkins <bsgcomp@arm.com>
Tested-by: Arm Jenkins <bsgcomp@arm.com>