feat: Provide libarm_compute.a in Linux(R) Cpu only multi-isa build via scons
Duplicate libarm_compute-static.a with name libarm_compute.a. libarm_compute-static.a will be removed in future releases, and libarm_compute.a will be used instead in all applicable releases.

This new binary will be accessible in build time only. It will be available in pre-built release binaries after it's been deprecated and rolled out to other builds with a major version upgrade.

Partially Resolves: COMPMID-8695

Change-Id: I59c3e809fc282d6dd16071184203e5be1d4af4d6
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>