test: Adjust the tolerance of fp16 depthwise convolution tests for cpu
Depthwise convolution tests for 5x5 filters were failing from time to time with small differences. The relative tolerance was %2, but running this test 10M times revealed the errors can be larger depending on how much the numbers are close to zero. Therefore, we added an absolute tolerance for the results that's been tested 10M times.

Resolves: COMPMID-8379
Change-Id: Iaa997b0208332bb39e3d30ad7991fc4627f9015a
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>