fix: Prevent potential overflow when building with clang-cl
This patch fixes a potential overflow raised by clang-cl when compiling CpuIsaInfo.cpp. The compiler uses 32-bits for unsigned long, and therefore 37-bit shift overflows. This variable is only defined in this context and not used. So, the fix doesn't impact any current problems, but in case the same variable is used somewhere else in the future where unsigned long is 32-bits or is not promoted to 64-bit integer in compile-time, this fix will ensure that it's treated as a 64-bit int.

Change-Id: I22c0d192f9664bb5f92e85a52ec4f57bcb32d829
Signed-off-by: Gunes Bayir <gunes.bayir@arm.com>